Update PR, following Matthieu's suggestions
